header cleanup
