+++ metagraph/src/graph/representation/hash/dbg_sshash.hpp
@@ -3,13 +3,9 @@
 
 #include <iostream>
 
-#include <tsl/ordered_set.h>
 #include <dictionary.hpp>
 
-#include "common/utils/string_utils.hpp"
-#include "common/logger.hpp"
 #include "graph/representation/base/sequence_graph.hpp"
-#include "kmer/kmer_extractor.hpp"
 
 namespace mtg::graph {
 
@@ -32,63 +28,6 @@ class IDictionary {
     virtual sshash::lookup_result lookup_advanced(
             char const* string_kmer,
             bool check_reverse_complement = true) const = 0;
-};
-
-template <typename KmerInt>
-class Dictionary : public IDictionary {
-  public:
-#if _PROTEIN_GRAPH
-    using kmer_t = sshash::aa_uint_kmer_t<KmerInt>;
-#else
-    using kmer_t = sshash::dna_uint_kmer_t<KmerInt>;
-#endif
-
-    const sshash::dictionary<kmer_t>& data() const { return dict_; }
-    size_t size() const { return dict_.size(); }
-
-    void visit(essentials::loader &loader) { dict_.visit(loader); }
-    void visit(essentials::saver &saver) const { dict_.visit(saver); }
-
-    void access(uint64_t kmer_id, char* string_kmer) const { dict_.access(kmer_id, string_kmer); }
-    void build(std::string const& input_filename, sshash::build_configuration const& build_config) {
-        dict_.build(input_filename, build_config);
-    }
-
-    uint64_t lookup(char const* string_kmer, bool check_reverse_complement = true) const {
-        return dict_.lookup(string_kmer, check_reverse_complement);
-    }
-
-    uint64_t lookup_uint(kmer_t uint_kmer, bool check_reverse_complement = true) const {
-        return dict_.lookup_uint(uint_kmer, check_reverse_complement);
-    }
-
-    sshash::lookup_result lookup_advanced(
-            char const* string_kmer,
-            bool check_reverse_complement = true) const {
-        return dict_.lookup_advanced(string_kmer, check_reverse_complement);
-    }
-
-    sshash::lookup_result lookup_advanced_uint(
-            kmer_t uint_kmer,
-            bool check_reverse_complement = true) const {
-        return dict_.lookup_advanced_uint(uint_kmer, check_reverse_complement);
-    }
-
-    sshash::neighbourhood<kmer_t> kmer_forward_neighbours(
-            char const* string_kmer,
-            bool check_reverse_complement = true) const {
-        return dict_.kmer_forward_neighbours(string_kmer, check_reverse_complement);
-    }
-
-    sshash::neighbourhood<kmer_t> kmer_backward_neighbours(
-            char const* string_kmer,
-            bool check_reverse_complement = true) const {
-        return dict_.kmer_backward_neighbours(string_kmer, check_reverse_complement);
-    }
-
-  private:
-    // TODO: this is a hack since sshash::dictionary<kmer_t>::visit is always non-const
-    mutable sshash::dictionary<kmer_t> dict_;
 };
 
 class DBGSSHash : public DeBruijnGraph {
